Move sharding-related native function tests to sharding.test.js.
